directory changes | added refetch functionality | delete functionality completed
