+++ components/incidents/list/components/NavigationBarMenuItem.tsx
@@ -7,7 +7,7 @@ import {
   spanInsideNav,
   navActiveItem,
   navNonActiveItem,
-} from "../styles/navStyles";
+} from "../overrides/navStyles";
 
 // components
 import { Block } from "baseui/block";
